requirements.txt
numpy==1.19.4
Flask==1.1.2
opencv-python==4.1.2.30
opencv-contrib-python==4.4.0.46
scipy==1.5.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
tensorflow==2.4.0
protobuf==3.12.4
scikit-image==0.16.2
imageio==2.4.1
pyflann-py3==0.1.0